Raw Data Files
pitchestagged.csv is the raw file in .csv format with the column "TaggedPitchType" filled in.
example.csv is the raw file in .csv format with the column "TaggedPitchType" not filled in, so the code uses the "AutoPitchType" from Trackman. Not as accurate, but still filters data appropriately given the change.
Dudzinski.xlsx is the excel saved file, which will be autoconverted to a .csv file which can be processed by the code.